requirements.txt
# Core
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4

# Market data
yfinance==0.2.54

# Sentiment analysis
vaderSentiment==3.3.2

# Scheduling / timezone
pytz==2024.1
